+++ SotatekCore/Util.swift
@@ -17,9 +17,9 @@ open class Util {
         return reachability.isReachable
     }
 
-    open static func currentTime() -> Int {
+    open static func currentTime() -> Int64 {
         let time = Date().timeIntervalSince1970
-        return Int(time * 1000)
+        return Int64(time * 1000)
     }
     
     open static func uuid() -> String {
